fix: align context window readouts
